+++ src/app/static/src/tests/root/mutations.test.ts
@@ -326,19 +326,4 @@ describe("Root mutations", () => {
         mutations.SetUpdatingLanguage(state, {payload: false});
         expect(state.updatingLanguage).toBe(false);
     });
-
-    it("can set ErrorReportError", () => {
-        const error = mockError("Error");
-        const state = mockRootState();
-        mutations.ErrorReportError(state, {payload: error});
-        expect(state.errorReportError).toBe(error);
-        expect(state.errorReportSuccess).toBe(false);
-    });
-
-    it("can set ErrorReportSuccess", () => {
-        const state = mockRootState();
-        mutations.ErrorReportSuccess(state);
-        expect(state.errorReportSuccess).toBe(true);
-        expect(state.errorReportError).toBe(null);
-    });
 });
